🐛 Break the import cycles of the shell package
CodeQL flagged that the registry, session, and store modules imported
one another. The exceptions move to a leaf module, the registry no
longer names the session type, and the app imports the commands
package explicitly instead of through an unused import.
