+++ bindings/mnt/pyfiction/cli/commands/verification.py
@@ -15,7 +15,8 @@ import io
 from typing import TYPE_CHECKING
 
 from mnt.pyfiction import eq_type, equivalence_checking, equivalence_checking_stats, gate_level_drvs, get_name
-from mnt.pyfiction.cli.registry import Category, CommandError, command, store_flags
+from mnt.pyfiction.cli.errors import CommandError
+from mnt.pyfiction.cli.registry import Category, command, store_flags
 from mnt.pyfiction.cli.session import stats_to_dict
 
 if TYPE_CHECKING:
